Safari-611.1.10/LayoutTests/http/tests/security/contentSecurityPolicy/connect-src-beacon-blocked-expected.txt

CONSOLE MESSAGE: Refused to connect to http://localhost:8000/security/contentSecurityPolicy/resources/echo-report.php because it does not appear in the connect-src directive of the Content Security Policy.
Pass
